fix : mypage
