+++ src/components/pages/Mypage/index.tsx
@@ -2,7 +2,6 @@ import { Routes, Route, Location, Navigate } from 'react-router-dom';
 import ManageRouter from '../Manage';
 import Mypage from './Mypage';
 import Enter from './Enter';
-import NewMypage from './NewMypage';
 import BackgroundTemplate from '@components/atoms/layout/BackgroundTemplate';
 import ForegroundTemplate from '@components/atoms/layout/ForegroundTemplate';
 
@@ -13,7 +12,7 @@ function MypageRouter({ location }: { location: Location }) {
         path="/"
         element={
           <BackgroundTemplate>
-            <NewMypage />
+            <Mypage />
           </BackgroundTemplate>
         }
       />
